requirements.txt
# Humbu Rural Bot Core - $147,575/month Government SaaS
# Core dependencies for revenue generation

# Web Framework
flask==2.3.3

# Hardware Interface
pyserial==3.5

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Configuration
pyyaml==6.0

# Logging
structlog==23.1.0

# API Documentation
flask-swagger-ui==4.11.1

# Testing
pytest==7.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Revenue Tracking
openpyxl==3.1.2  # For revenue reports

# Hardware Specific
RPi.GPIO==0.7.1  # For Raspberry Pi compatibility
